requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
jupyter>=1.0.0